+++ src/identity/ed25519/index.ts
@@ -1,1 +1,1 @@
-export * from "./ed25519-key-pair-user-identity"
+export * from "./ed25519-key-pair-identity"
